backend/requirements.txt
fastapi
uvicorn
pydantic
email-validator
sqlalchemy
pymysql
python-dotenv
pymongo
